Fixed bugs in the table generation, which was reversed and caused NPE on some cases. Also changed ints to longs to avoid clipping which caused bad predictor behaviour

+++ src/metrics.c
@@ -1,6 +1,6 @@
 #include "metrics.h"
 
-double mse(int * data1, int * data2, int length) {
+double mse(long * data1, long * data2, long length) {
 	double acc = 0;
 	for (int i = 0; i < length; i++) {
 		acc += (data1[i] - data2[i])*(data1[i] - data2[i]);
