+++ src/metrics.h
@@ -1,6 +1,6 @@
 #ifndef __METRICS_H__
 #define __METRICS_H__
 
-double mse(int * data1, int * data2, int length);
+double mse(long * data1, long * data2, long length);
 
 #endif //__METRICS_H__
